Apply consistent styling for changed resource nodes
This refactors the logic so that we have a common "resourceNode" interface, which then
is implemented for both the regular resource nodes (internally *cgraph.Nodes) and the
"fake" nodes we use for nested stacks (internally *cgraph.Graphs).

By doing that we can use the same styling logic for either, which makes removal of nested
stacks visible.

Fixes: https://github.com/ankon/explain-cloudformation-changeset/issues/7