Generate edges for all changes
graphviz/cgraph allows to have multiple edges between the same node pair in a non-strict
graph, as long as the edges have distinct names. Unfortunately the version of graphviz go-graphviz
embeds has a bug in exactly the logic for looking up the edges, so this commit also switches
out the version to a fork that has a suitable fix for the problem.

See https://gitlab.com/graphviz/graphviz/-/issues/1191 for the original upstream problem, and
https://github.com/goccy/go-graphviz/issues/52 for the issue in go-graphviz.

Fixes https://github.com/ankon/explain-cloudformation-changeset/issues/6