Use a matrix build for releases
We cannot do a "simple cross-build" for our releases, because we have native library
code in the graphviz library.

Change the logic to run a matrix; in development mode only on ubuntu-latest, when releasing
also on macos-latest. The release build then picks up the artifact, renames it, and attaches
it to the proper place.